Opt-in to coverage using spec:coverage. Configure coverage in the spec_helper because it has to happen before the app is loaded.
