+++ lib/raygun/app_builder.rb
@@ -271,6 +271,39 @@ RUBY
     def setup_simplecov
       copy_file '_lib/tasks/coverage.rake', 'lib/tasks/coverage.rake'
       copy_file '_spec/support/simplecov.rb', 'spec/support/simplecov.rb'
+
+
+      coverage = <<-RUBY
+# Coverage must be enabled before the application is loaded.
+if ENV['COVERAGE']
+  require 'simplecov'
+
+  # Writes the coverage stat to a file to be used by Cane.
+  class SimpleCov::Formatter::QualityFormatter
+    def format(result)
+      SimpleCov::Formatter::HTMLFormatter.new.format(result)
+      File.open('coverage/covered_percent', 'w') do |f|
+        f.puts result.source_files.covered_percent.to_f
+      end
+    end
+  end
+  SimpleCov.formatter = SimpleCov::Formatter::QualityFormatter
+
+  SimpleCov.start do
+    add_filter '/spec/'
+    add_filter '/config/'
+    add_filter '/vendor/'
+    add_group  'Models', 'app/models'
+    add_group  'Controllers', 'app/controllers'
+    add_group  'Helpers', 'app/helpers'
+    add_group  'Views', 'app/views'
+    add_group  'Mailers', 'app/mailers'
+  end
+end
+
+RUBY
+
+      prepend_to_file 'spec/spec_helper.rb', coverage
     end
 
     def setup_default_rake_task
