+++ templates/_lib/tasks/coverage.rake
@@ -3,10 +3,8 @@ namespace :spec do
     ENV['COVERAGE'] = '1'
   end
 
-  desc 'Executes specs with code coverage reports'
+  desc "Executes specs with code coverage reports"
   task coverage: :enable_coverage do
     Rake::Task[:spec].invoke
   end
 end
-
-Rake::Task['spec'].enhance ['spec:enable_coverage']
